terceiro dia

+++ src/routes/memories.ts
@@ -4,8 +4,13 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-   app.get('/memories', async () => {
+   app.addHook('preHandler', async (request) => {
+      await request.jwtVerify()
+   })
+
+   app.get('/memories', async (request) => {
       const memories = await prisma.memory.findMany({
+         where: { userId: request.user.sub },
          orderBy: { createdAt: 'asc' },
       })
 
@@ -18,7 +23,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
       })
    })
 
-   app.get('/memories/:id', async (request) => {
+   app.get('/memories/:id', async (request, reply) => {
       const paramsSchema = z.object({
          id: z.string().uuid(),
       })
@@ -26,6 +31,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
       const { id } = paramsSchema.parse(request.params)
 
       const memory = await prisma.memory.findUniqueOrThrow({ where: { id } })
+
+      if (!memory.isPublic && memory.userId !== request.user.sub)
+         return reply.status(401).send()
 
       return memory
    })
@@ -44,14 +52,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
             content,
             coverUrl,
             isPublic,
-            userId: '235fec9d-1ab5-4019-86dd-e2b33cb04c24',
+            userId: request.user.sub,
          },
       })
 
       return memory
    })
 
-   app.put('/memories/:id', async (request) => {
+   app.put('/memories/:id', async (request, reply) => {
       const paramsSchema = z.object({
          id: z.string().uuid(),
       })
@@ -66,7 +74,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
       const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-      const memory = await prisma.memory.update({
+      let memory = await prisma.memory.findUniqueOrThrow({ where: { id } })
+
+      if (memory.userId !== request.user.sub) return reply.status(401).send()
+
+      memory = await prisma.memory.update({
          where: { id },
          data: { content, coverUrl, isPublic },
       })
@@ -74,12 +86,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
       return memory
    })
 
-   app.delete('/memories/:id', async (request) => {
+   app.delete('/memories/:id', async (request, reply) => {
       const paramsSchema = z.object({
          id: z.string().uuid(),
       })
 
       const { id } = paramsSchema.parse(request.params)
+
+      const memory = await prisma.memory.findUniqueOrThrow({ where: { id } })
+
+      if (memory.userId !== request.user.sub) return reply.status(401).send()
 
       await prisma.memory.delete({ where: { id } })
    })
